Assignment 5
Task #1: (6 pts) Create a function scrape_data(start_date, from_place, to_place, city_name). This function takes 4 arguments:

start_date: this is a datetime object for the start date that you should use in your query to Google Flight explorer. Note: this is the d param in this URL (Links to an external site.). You'll see for this query that the start date is 2017-03-25.
from_place: this is a string with the name of the origin of the flights, e.g. . This is what you would type into the From field on the Flight Explorer page. NOTE: if you choose not to use Selenium's send_keys to put this data in the From field similar to what I did in class, BE SURE that you are searching over all the airports that belong to the from place. For instance, if you type NYC into the From field on the flight explore page and hit Enter, notice that it's searching over all airports in the NYC area: JFK, LGA, EWR. Make sure yours does the same. DON'T ASSUME THAT I WILL USE NYC. I will check to make sure your results